+++ crates/cli/src/commands/install_all.rs
@@ -1,7 +1,7 @@
 use crate::commands::install::install;
 use crate::helpers::enable_logging;
 use crate::tools::ToolType;
-use proto_core::{ProtoError, ToolsConfig, CONFIG_NAME};
+use proto_core::{ProtoError, ToolsConfig, TOOLS_CONFIG_NAME};
 use std::{env, str::FromStr};
 
 pub async fn install_all() -> Result<(), ProtoError> {
@@ -10,7 +10,7 @@ pub async fn install_all() -> Result<(), ProtoError> {
     let current_dir = env::current_dir().expect("Invalid working directory!");
 
     let Some(config) = ToolsConfig::load_upwards(&current_dir)? else {
-        return Err(ProtoError::MissingConfig(CONFIG_NAME.to_owned()));
+        return Err(ProtoError::MissingConfig(TOOLS_CONFIG_NAME.to_owned()));
     };
 
     let mut futures = vec![];