+++ crates/core/src/detector.rs
@@ -3,7 +3,7 @@
 use crate::errors::ProtoError;
 use crate::manifest::{Manifest, MANIFEST_NAME};
 use crate::tool::Tool;
-use crate::tools_config::{ToolsConfig, CONFIG_NAME};
+use crate::tools_config::{ToolsConfig, TOOLS_CONFIG_NAME};
 use crate::{color, is_version_alias};
 use lenient_semver::Version;
 use log::{debug, trace};
@@ -76,7 +76,7 @@ pub async fn detect_version_from_environment<'l, T: Tool<'l> + ?Sized>(
             trace!(
                 target: "proto:detector",
                 "Checking proto configuration file ({})",
-                CONFIG_NAME
+                TOOLS_CONFIG_NAME
             );
 
             let config = ToolsConfig::load_from(dir)?;